# doc/api-documentation.yml
---
openapi: 3.0.0
info:
  version: 2.2.0
  title: HCGateway
  description: ''
servers:
- url: https://api.hcgateway.shuchir.dev
paths:
  "/api/v2/login":
    post:
      tags:
        - Authentication
      operationId: get_token
      description: Login a user and get a token to use for authentiction.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/loginParameters"
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/loginObject"
  "/api/v2/refresh":
    post:
      tags:
        - Authentication
      operationId: refresh_token
      description: Refresh a user token.
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                refresh:
                  nullable: false
                  type: string
                  description: The refresh token you got when you logged in.
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/loginObject"
  "/api/v2/revoke":
    post:
      tags:
        - Authentication
      operationId: revoke_token
      description: Revoke all user tokens and refresh tokens immediately. Pass a currently valid token through the Authentication header to revoke your user.
      parameters: []
      security:
      - bearerAuth: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/silentObject"
  "/api/v2/fetch/{method}":
    post:
      operationId: fetch
      description: Get data from the database.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: method
        required: true
        schema:
          type: string
        description: 'Can be one of the following: activeCaloriesBurned, basalBodyTemperature,
          basalMetabolicRate, bloodGlucose, bloodPressure, bodyFat, bodyTemperature,
          boneMass, cervicalMucus, distance, exerciseSession, elevationGained, floorsClimbed,
          heartRate, height, hydration, leanBodyMass, menstruationFlow, menstruationPeriod,
          nutrition, ovulationTest, oxygenSaturation, power, respiratoryRate, restingHeartRate,
          sleepSession, speed, steps, stepsCadence, totalCaloriesBurned, vo2Max, weight,
          wheelchairPushes'
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/fetchParameters"
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/dataObject"
  "/api/v2/push/{method}":
    put:
      operationId: pushData
      description: Insert data into your local Health Connect store. This method does
        not add the data to the HCGateway database. This method is only used to push
        data to your local Health Connect store. This is so you can sync + fetch to
        verify the record has been added.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: method
        required: true
        schema:
          type: string
        description: 'Can be one of the following: activeCaloriesBurned, basalBodyTemperature,
          basalMetabolicRate, bloodGlucose, bloodPressure, bodyFat, bodyTemperature,
          boneMass, cervicalMucus, distance, exerciseSession, elevationGained, floorsClimbed,
          heartRate, height, hydration, leanBodyMass, menstruationFlow, menstruationPeriod,
          nutrition, ovulationTest, oxygenSaturation, power, respiratoryRate, restingHeartRate,
          sleepSession, speed, steps, stepsCadence, totalCaloriesBurned, vo2Max, weight,
          wheelchairPushes'
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/pushParameters"
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/silentObject"
  "/api/v2/delete/{method}":
    delete:
      operationId: deleteData
      description: Request deletion for a record. This method does not delete the
        record from the HCGateway database. This method is only used to request deletion
        of a record from your local Health Connect store. The app will delete it from
        the database using the sync method once it has deleted the record(s) locally.
      security:
      - bearerAuth: []
      parameters:
      - in: path
        name: method
        required: true
        schema:
          type: string
        description: 'Can be one of the following: activeCaloriesBurned, basalBodyTemperature,
          basalMetabolicRate, bloodGlucose, bloodPressure, bodyFat, bodyTemperature,
          boneMass, cervicalMucus, distance, exerciseSession, elevationGained, floorsClimbed,
          heartRate, height, hydration, leanBodyMass, menstruationFlow, menstruationPeriod,
          nutrition, ovulationTest, oxygenSaturation, power, respiratoryRate, restingHeartRate,
          sleepSession, speed, steps, stepsCadence, totalCaloriesBurned, vo2Max, weight,
          wheelchairPushes'
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/deleteParameters"
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/silentObject"
  "/api/v2/analytics/snapshot":
    get:
      tags: [Analytics]
      operationId: getAnalyticsSnapshot
      description: Return the current frontend-ready health analytics snapshot.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: The current immutable analytics run.
          content:
            application/json:
              schema:
                type: object
                required: [generatedAt, source, sleepSessions, analytics]
                properties:
                  generatedAt: {type: string, format: date-time}
                  source: {type: string, example: health-connect}
                  sleepSessions: {type: array, items: {type: object}}
                  analytics: {type: object}
        '404': {description: The first background analytics run is not ready.}
  "/api/v2/sync/status":
    get:
      tags: [Sync]
      operationId: getPhoneSyncStatus
      description: >-
        Return server-observed phone upload activity. observedActive is true for
        120 seconds after the latest authenticated sync upload; it does not claim
        that the Android foreground task itself is still running.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Phone upload activity and lifetime counters.
          content:
            application/json:
              schema:
                type: object
                required: [observedActive, state, activityWindowSeconds]
                properties:
                  observedActive: {type: boolean}
                  state: {type: string, enum: [receiving, idle, never_observed]}
                  lastUploadAt: {type: string, format: date-time, nullable: true}
                  activeUntil: {type: string, format: date-time, nullable: true}
                  secondsSinceLastUpload: {type: integer, nullable: true}
                  lastRecordType: {type: string, nullable: true}
                  lastRecordCount: {type: integer, nullable: true}
                  totalUploadRequests: {type: integer}
                  totalRecordsReceived: {type: integer}
                  activityWindowSeconds: {type: integer, example: 120}
                  note: {type: string}
  "/api/v2/analytics/daily":
    get:
      tags: [Analytics]
      operationId: getDailyAnalytics
      description: Return date-indexed documents from the current analytics run.
      security:
      - bearerAuth: []
      parameters:
      - {in: query, name: start, schema: {type: string, format: date}}
      - {in: query, name: end, schema: {type: string, format: date}}
      - {in: query, name: limit, schema: {type: integer, minimum: 1, maximum: 1000, default: 400}}
      responses:
        '200':
          description: Daily prepared analytics.
          content:
            application/json:
              schema:
                type: object
                properties:
                  runId: {type: string}
                  count: {type: integer}
                  days: {type: array, items: {type: object}}
  "/api/v2/analytics/day":
    get:
      tags: [Analytics]
      operationId: getHealthDayDashboard
      description: >-
        Return the frontend-oriented health day contract plus nearby-day summaries.
        Every requested metric includes an explicit availability status and missing-data note.
      security:
      - bearerAuth: []
      parameters:
      - {in: query, name: date, schema: {type: string, format: date}, description: Defaults to today in the configured home timezone.}
      - {in: query, name: radius, schema: {type: integer, minimum: 0, maximum: 7, default: 7}}
      responses:
        '200':
          description: A shaped day view and low-resolution nearby days.
          content:
            application/json:
              schema:
                type: object
                required: [contractVersion, day, nearbyDays, runId]
                properties:
                  contractVersion: {type: string, example: health-day-v1}
                  day: {type: object}
                  nearbyDays: {type: array, items: {type: object}}
                  runId: {type: string}
        '404': {description: The first background analytics run is not ready.}
  "/api/v2/analytics/status":
    get:
      tags: [Analytics]
      operationId: getAnalyticsStatus
      description: Return analytics job/current-run metadata and the same server-observed phoneSync status.
      security:
      - bearerAuth: []
      responses:
        '200': {description: Analytics status.}
  "/api/v2/analytics/rebuild":
    post:
      tags: [Analytics]
      operationId: rebuildAnalytics
      description: Queue a non-blocking analytics rebuild.
      security:
      - bearerAuth: []
      responses:
        '202': {description: Rebuild queued.}
  "/api/v2/analytics/config":
    get:
      tags: [Analytics]
      operationId: getAnalyticsConfig
      security:
      - bearerAuth: []
      responses:
        '200': {description: Effective analytics configuration.}
    put:
      tags: [Analytics]
      operationId: updateAnalyticsConfig
      description: Update analytics configuration and queue a rebuild.
      security:
      - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                homeTimeZone: {type: string, example: America/Los_Angeles}
                sleepTargetMinutes: {type: integer, minimum: 240, maximum: 720}
                birthDate: {type: string, format: date, nullable: true}
                heartRateZoneThresholds:
                  type: array
                  minItems: 6
                  maxItems: 6
                  items: {type: number, minimum: 30, maximum: 240}
                  description: Increasing BPM thresholds for zone 1 through zone 5 and personal maximum.
                heartRateZoneTestDate: {type: string, format: date, nullable: true}
      responses:
        '202': {description: Configuration accepted and rebuild queued.}
  "/api/v2/analytics/inventory":
    get:
      tags: [Analytics]
      operationId: getAnalyticsInventory
      description: Return raw signal, date-range, and source counts without decrypting values.
      security:
      - bearerAuth: []
      responses:
        '200': {description: Raw-data inventory.}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    dataObject:
      properties:
        _id:
          nullable: false
          type: string
          description: The id Health Connect stored the object with.
        data:
          nullable: false
          type: object
          description: The actual data stored in Health Connect
        id:
          nullable: false
          type: string
        start:
          nullable: false
          type: string
          format: date-time
          description: The start time of the data
        end:
          nullable: true
          type: string
          format: date-time
          description: The end time of the data. Only present if the data is a time-series.
        app:
          nullable: false
          type: string
          description: The app that created the data
    silentObject:
      properties:
        success:
          nullable: false
          type: boolean
          description: Sucessfully sent the data or not
        message:
          nullable: false
          type: string
          description: A confirmation message.
    loginObject:
      properties:
        token:
          nullable: false
          type: string
          description: The bearer token you should use for future requests.
        refresh:
          nullable: false
          type: string
          description: The refresh token you should use to get a new token after the current one expires.
        expiry:
          nullable: false
          type: string
          format: date-time
          description: The time the token expires. This is always 12 hours after the token was issued.
    loginParameters:
      properties:
        username:
          nullable: false
          type: string
          required: true
        password:
          nullable: false
          type: string
          required: true
    fetchParameters:
      properties:
        queries:
          nullable: true
          type: object
          description: Queries to use to filter the data. Must be a valid mongodb
            query. See https://www.mongodb.com/docs/languages/python/pymongo-driver/current/read/specify-a-query/
            for more information.
    pushParameters:
      properties:
        data:
          nullable: false
          type: array
          items:
            type: object
          description: The record to push. This can be an object or array of objects.
            See https://developer.android.com/reference/kotlin/androidx/health/connect/client/records/package-summary#classes
            for each data type's object structure.
    deleteParameters:
      properties:
        uuid:
          nullable: false
          type: array
          items:
            type: string
          description: The uuid of the record to delete. Can be a string or an array
            of strings.